Reorganize files and class hierarchy
Use puppet's automatic loading mechanism. When using 'import
something/*' you need to restart the puppetmaster when a file is added
else it will not be seen. Plus, it's way easier to find things this way:
just follow the class/define name.

I crammed in there some changes in hierarchy to add 'voicemail' under
account/ since it makes more sense there and change 'context' to be under
extensions/ so that the define has a more explicit name.
I also got rid of the with- prefixes since the hierarchy already
describes the addition clearly enough.