manage files that include files in .d dirs.
Instead of simply managing the "Include" line, we'll manage the whole
file.

Since we're gonna replace content with those files, we'll rename them to
the same thing as what they'll be replacing.

Add missing voicemail.conf file.

Note that for now they're static files, but they should very soon be
transformed into templates so that different values in the "general"
context can be customized.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>