Require and notify links are wrong
Currently, it's impossible to install asterisk because require and
notify links refer directly to package and service resouces.

First case: the main class already forces an order so inside install,
config and service classes those links are unnecessary.

Second case: for deeper-nested defines, we should require and notify the
install and the service classes instead of referencing things that are
done inside them. It simplifies links and ensures that we always refer
to the right thing even if things change inside those classes.

Signed-off-by: Gabriel Filion <gabster@lelutin.ca>