Simplify iax and sip templates.
The comments that come with the default config in debian are great, but
they are useless in the context of a file managed by puppet.

Plus, they force us to special-case every darned option.

Let's stay simple and KISS all that cruft goodbye.

Note: the options from the hash are sorted so that the contents of the
files will not change on every puppet run.

Signed-off-by: Gabriel Filion <gabster@lelutin.ca>