Collapse all configuration into asterisk::config
All of the classes that do only one thing, calling config_dotd, are
actually just configuration that should be done for the rest of the
module to work OK.

So let's put all of that stuff inside the config class and get rid of
those other classes. It'll centralize basic configuration and clean up
the module by reducing the number of files.

Signed-off-by: Gabriel Filion <gabster@lelutin.ca>