SIP/IAX options should only override default values.
For users, having to provide the hole default hash with just a couple of
changes is cumbersome. It's better for usability if the options hash
parameters would only override default values or set new ones.

Now that we override values, we need to ensure that both main class
parameters are actually hashes, since we're passing them directly to the
merge() function from stdlib that requires values to be hashes.

Signed-off-by: Gabriel Filion <gabster@lelutin.ca>