Manage modules.conf as a template
This file doesn't need a .d directory since we can specify everything in
it via main class parameters.

Having a .d dir might arguably be interesting to enable modules in
arbitrary manifests without having to specify that value to the main
class (which might be problematic); we might be interested in adding
this in the future if requested/needed.

Signed-off-by: Gabriel Filion <gabster@lelutin.ca>