Make null files optional.
Some users might find it confusing to see a file called "null.cfg" in .d
dirs and might want to avoid placing such a file there. This file is
there for a good reason, but is only useful when the .d directory is
empty.

We'll keep the big bad ugly comment that explains why this file is
useful though.