Remove unfunctional class unimrcp
This class fails to prove itself useful since it uses a package that is
not in debian repositories. In order to use unimrcp, one would need to
add an additional package source, and doing so is out of the scope of
this module.

If people need to use this, they can define the apt source and install
this package at a higher level.