README: rework section hierarchy and split out extra features
For easier reading we'll use more first level sections.

The extra features section is useless and hides some elements of
configuration. Move subsections of this up one level, into Configuring
Asterisk directly.

Rename "Notes on Options" to something more descriptive and make it a
sub-section of "Parameters to the asterisk class".