use data types for the rest of the code
the previous attempt at moving towards puppet 4 data types was
incomplete. let's continue this switch with all missing resources. This
helps to considerably diminish the number of lines of code since type
checking is now handled for us by puppet itself.

This series of changes introduces a good number of Sensitive resources
for passwords and will require users of this module to change how they
call this module's resources.