use puppet-strings format to document all classes and defines
This should make it easier for users to find out what is happening. This
module has lots of options and it's nice to be able to find out what
they are.

I've decided to list out all options instead of deferring to URLs.
Documentation for asterisk is hard to find and usually outdated, so it's
useful to have information from multiple different sources collected as
documentation here.