Use some more precise data types and identify files that need the same
The files that only get identified need much more work. This commit is
mostly changing the files where it's easy to see where improvements can
be made.

Stdlib has some types that could become really handy. We're using some
of them here to restrict the format of some values.

In voicemail, favor a sound interface rather than simplicity to output
data in the template: marking data as undef is a lot clearer that it
means it's absent.