rename sip param calllimit to call_limit
this parameter is output as the option "call-limit". I usually reuse the
asterisk option name when all words are glued together (to make it
easier to know what it corresponds to in configuration) but doing so
when the option has a dash in its name just makes things readable for
no reason.