permit interleave of permit and deny for sip
to do this, we merge both permit and deny parameters into one list. each
element is either a permit or a deny rule and it will get output to
configuration in the same order as insertion happened in the list.

This change should let users configure more complex permissions for
peers.