keep entries in insert order for all config files
This is a continuation of the previous commit. Asterisk has some options
that are order-dependant, so using sort prevents users from representing
certain complex configurations.

The use of "sort" was previously inserted to avoid instabilities in the
contents of config files since puppet used to have no garantees with
regards to ordering of items in arrays and hashes. This is no longer the
case, and puppet now ensures that insertion order is kept, so we can
stop using sort in all configuration files.