Rename logger_options to logger_general and minimize logger.conf
The _options suffix was present when archii added the feature to their
repository but the asterisk module moved to a different naming scheme
for options that makes it easier to know where things land in the main
file. So let's follow this pattern.

The logger.conf was copied as-is from the default upstream file,
however this module currently tends to minimize the main file's
contents. It makes it easier to read the results.

Finally, let's also fix some tiny lint problems.