Add param to main class to manage log files
The file logger.conf was being managed by the previous changes but it
was impossible to change the log file configurations that were already
present or to add new ones.

With this new parameter, users are able to configure log files and also
to override the default configurations.

Log file configuration is really just a simple line for each log file,
so we don't need to use a .d directory for this.