Bump puppet an OS versions. Drop EOL versions
CI currently can't run tests against puppet 4 and 5 and they are both
EOL so let's drop support for those versions. We should start testing
against more recent versions of puppet however.

Debian 8 and 9 are EOL. Debian 10 is EOL but it is currently still in
LTS. Debian 11 is the current stable release.