Remove augeas resource
On newer debian releases with systemd, the file /etc/default/asterisk is
not used anymore.

It's possible that RHEL still needs this but I'll fix it eventually.

augeas resources were split out from puppet core, so to use this I'd
have to declare another dependency to the module which would be
ridiculous for only one resource.