Use only one template for all config files
This has been on my mind for a long time. All config files in asterisk
are very similar. They're kind of ini-ish but not exactly. But the
differences between the files are very minor. So keeping around one
template for each different asterisk configuration file adds more files
that need to be managed individually.

This deduplicates what's used for generating configuration.

This change also replaces old .erb files with a .epp template.

It also changes some tiny details:

 * includes for config files inside .d directories is now relative
   instead of absolute, and they force the .conf extension
 * directories and config files are now owned by the asterisk user. I
   have no idea why they were set to be owned by root in the first
   place. This change minimizes the number of change lines when puppet
   first configures the service from scratch.