Remove option agents_multiplelogin
What even is that option?

I think that when I added management of agents.conf I found this option
in
http://www.asteriskdocs.org/en/3rd_Edition/asterisk-book-html-chunk/ACD_id288918.html

But the current upstream documentation says that the general section is
not used and is reserved for future use. (although weirdly enough it's
talking about the "global" section whereas the sample config file that's
shipped with asterisk mentions this for the "general" section --
asterisk documentation apparently still remains very bad)
https://wiki.asterisk.org/wiki/display/AST/Asterisk+20+Configuration_app_agent_pool

So this parameter has no reason to exist.