Add parameter for purging unknown configuration files
With this new parameter you can make sure that the state of your
asterisk in known and corresponds to what you've setup in puppet.

It is set to false by default since doing the inverse could be
disruptive or downright destructive for some users.

The confdir is also getting its permissions managed in all cases now.