Fix or ignore all issues reported by puppet-lint
The new stdlib types that I've used make the types' intents explicit but
they require us to push the required stdlib version to a much more
recent version -- especially because of Stdlib::CreateResources, but
that type will make the REFERENCE file that much easier to grok.

I haven't checked yet whether this push to stdlib is problematic to the
supported puppet versions.