Let bundler figure out dependencies
ruby gems have metadata on the minimum supported ruby version and
supported versions of other libs so it can very well discover
limitations by itself without us hand-holding it.

also, imposing an upper bound on versions is an anti-pattern IMHO so
let's just stop that. if something breaks for real when new versions are
released, then we can add an upper bound on version dependencies *for a
reason*.